Create /src folder and move files to it
