+++ src/index.ts
@@ -1,0 +1,5 @@
+import {AppRegistry} from 'react-native';
+import App from './components/App';
+import {name as appName} from '../app.json';
+
+AppRegistry.registerComponent(appName, () => App);
